requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anthropic==0.33.1
anyio==4.4.0
blinker==1.8.2
boto3==1.34.159
botocore==1.34.159
cachetools==5.4.0
certifi==2024.7.4
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.15.4
filetype==1.2.0
fire==0.6.0
Flask==3.0.3
fonttools==4.53.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.65.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
openai==1.40.6
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybboxes==0.1.6
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.37
rsa==4.9
s3transfer==0.10.2
sahi==0.11.18
scipy==1.14.0
seaborn==0.13.2
shapely==2.0.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sympy==1.13.2
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
tokenizers==0.20.0
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.76
ultralytics-thop==2.0.0
urllib3==2.2.2
Werkzeug==3.0.3
yolov5==7.0.13
